上传图片完成。

+++ model/format.go
@@ -1,6 +1,6 @@
 package model
 
 type Format struct {
-	Id   int    `json:"id"`
+	Id   uint64 `json:"id"`
 	Name string `json:"name",omitempty"`
 }
